server_dev/requirements.txt
asgiref==3.2.7
Django==3.0.5
pytz==2019.3
sqlparse==0.3.1
uwsgi
Pillow
django-phonenumber-field[phonenumbers]
django-phonenumber-field[phonenumberslite]